Ferment a brew

Missing values can occur in the training data, testing data, and validation data. An important requirement for missing value strategies is that only information from the training data should be used to impute missing data.

Unfortunately, some imputation strategies are not designed to work this way! For example, softImpute imputes missing values based on the index of the missing value in the training data, and this doesn't generalize to testing data because testing data (by definition) are not in the training data.

Fortunately, ferment gives you options. For example, you can use softImpute to impute missing training data, and then use kneighbors to impute the testing data by matching the testing observations to their nearest neighbors in the training data. More specifically, you can tell ferment to impute the testing data using either the original unimputed training data or using each of the imputed training datasets, separately. The same options apply for all ipa_brew objects.

ferment(brew, ..., dbl_impute = FALSE, dbl_neighbors = 5)

# S3 method for softImpute_brew
ferment(brew, ..., dbl_impute = FALSE,
  dbl_neighbors = 5)

# S3 method for kneighbors_brew
ferment(brew, ..., dbl_impute = FALSE,
  dbl_neighbors = 5)

# S3 method for missRanger_brew
ferment(brew, ..., dbl_impute = FALSE,
  dbl_neighbors = 5)

is_fermented(brew)

Arguments

brew
an ipa_brew object.

...
Name-value pairs of datasets that will be imputed. The name of each argument will be the name of the variable in the wort of the ipa_brew that comprises imputed datasets. The value of each argument will be a dataset that is imputed and then placed into the designated column in the wort.

dbl_impute
(TRUE/FALSE). If TRUE, then imputed training data will be used to impute missing values in new_data. Otherwise, the original data (i.e., the data containing missing values) will be used to impute missing values in new_data.

dbl_neighbors
an integer value specifying the number of nearest neighbors to identify when imputing testing data. This argument only applies when the given ipa_brew does not have a known method to impute new data. If you are making a kneighbors brew, this argument is only relevant when dbl_impute = FALSE. Otherwise, the sequence of neighbors used to create imputed training sets will be copied when imputing testing data.

Methods (by class)

softImpute_brew: impute missing values for softImpute_brew objects.

kneighbors_brew: impute missing values for kneighbors_brew objects.

missRanger_brew: impute missing values for missRanger_brew objects.

Note

What is a wort? A component of a brew object that contains imputed datasets, models used to impute those datasets, and the corresponding hyper-parameters of those models.

Examples

data <- data.frame(
  x1 = 1:10,
  x2 = 10:1,
  x3 = rnorm(10),
  outcome = 11 + runif(10)
)

data[1:2, 1] = NA
data[5:6, 2] = NA

new_data = data.frame(
  x1 = 1:10,
  x2 = 10:1,
  x3 = rnorm(10),
  outcome = 11 + runif(10)
)

new_data[6:7, 1] = NA
new_data[2:3, 2] = NA

knn_brew <- brew(data, outcome = outcome, flavor = 'kneighbors')

spicy_knn <- spice(
  brew = knn_brew,
  neighbors = c(3, 5),
  aggr_neighbors = c(TRUE, FALSE)
)

mashed_knn <- mash(spicy_knn)

fermented_knn <- ferment(mashed_knn, new_data = new_data)

set.seed(101)
n=200
p=100
J=50
np=n*p
missfrac=0.3
x=matrix(rnorm(n*J),n,J)%*%matrix(rnorm(J*p),J,p)+matrix(rnorm(np),n,p)/5
ix=seq(np)
imiss=sample(ix,np*missfrac,replace=FALSE)
xna=x
xna[imiss]=NA

data <- as.data.frame(xna)

soft_brew <- brew(data, outcome = c(V1,V2), flavor = 'softImpute')
soft_brew <- spice(soft_brew, n_impute = 10, step_size = 2)
soft_brew <- mash(soft_brew, scale_lambda = 0.12)