a simple framework to simulate simple dataframes

Simulated data allow analysts to conduct controlled experiments using specific parameters to generate data. This function provides a general framework to simulate missing data under different patterns:

gen_simdata(problem_type = c("regression", "classification", "survival"),
  ncov = 3, nint = 2, degree = 3, rho = 1/2, corstr = c("AR1",
  "CS"), nobs = 10000, error_sd = 1/2, prevalence = NULL,
  split_prop = 1/2, miss_pattern = "mar", npatterns = 10,
  trn_miss_prop = 1/2, tst_miss_prop = 0)

Arguments

problem_type
A character value indicating the problem type to simulate data for. Valid options are 'regression', 'classification', or 'survival'.

ncov
the number of main effects used to generate an outcome variable

nint
the number of interaction effects used to generate an outcome variable.

degree
the degree of each predictor variable's relationship to the outcome. For example, degree = 2 makes the relationship between each predictor variable and the outcome quadratic.

rho
the correlation constant among predictors in the X matrix.

corstr
The correlation structure among predictors in the X matrix.

nobs
the total number of observations in the simulated data.

error_sd
the standard deviation of error applied when generating outcome values.

prevalence
(this is only relevant for classification problems). The prevalence of the outcome.

split_prop
the proportion of data that will be randomly assigned to the training dataset.

miss_pattern
A character value indicating what type of missingness pattern to apply. Valid options are 'mcar', 'mar', and 'mnar'. mcar = missing completely at random, mar = missing at random, and mnar = missing not at random.

npatterns
The number of missing patterns used to ampute data.

trn_miss_prop
The proportion of data in the training set that will be set to missing.

tst_miss_prop
The proportion of data in the testing set that will be set to missing.

Details

missing completely at random (MCAR): missingness occurs at random, independent of all variables

missing at random (MAR): missingness occurs at random, conditional on measured variables in the observed data.

missing not at random (MNAR) missingess occurs non-randomly and is dependent on unmeasured variables.

All simulated predictor variables are numeric. Regression coefficients are generated randomly (all values are between -1 and 1).