Mash a brew

When an ipa_brew is mashed, training data are used to fit imputation models based on the brew flavor.

mash(brew, with = NULL, ...)

# S3 method for softImpute_brew
mash(brew, with = NULL, ...)

# S3 method for kneighbors_brew
mash(brew, with = NULL, ...)

# S3 method for missRanger_brew
mash(brew, with = NULL, ...)

is_mashed(brew)

Arguments

brew
an ipa_brew object that may or may not have been spiced (see spice).

with
the output of a helper function for mashing brews. The helper functions are masher_nbrs, masher_soft, and masher_rngr.

...
additional arguments for specific brew flavors.

Methods (by class)

softImpute_brew: Mash a soft brew (use masher_soft for with).

kneighbors_brew: Mash a neighbor's brew (use masher_nbrs for with)

missRanger_brew: Mash a ranger's brew (use masher_rngr for with)

Examples

set.seed(101)
n=200
p=100
J=50
np=n*p
missfrac=0.3
x=matrix(rnorm(n*J),n,J)%*%matrix(rnorm(J*p),J,p)+matrix(rnorm(np),n,p)/5
ix=seq(np)
imiss=sample(ix,np*missfrac,replace=FALSE)
xna=x
xna[imiss]=NA

data <- as.data.frame(xna)

soft_brew <- brew(data, outcome = V1, flavor = 'softImpute')
soft_brew <- spice(soft_brew, n_impute = 10, step_size = 2)
soft_brew <- mash(soft_brew, scale_lambda = 0.12)

soft_brew$wort

#> # A tibble: 10 x 4
#>    impute lambda  rank fit             
#>     <int>  <dbl> <int> <list>          
#>  1      1  23.6      3 <named list [3]>
#>  2      2  21.1      5 <named list [3]>
#>  3      3  18.6      7 <named list [3]>
#>  4      4  16.1      9 <named list [3]>
#>  5      5  13.6     11 <named list [3]>
#>  6      6  11.0     13 <named list [3]>
#>  7      7   8.54    15 <named list [3]>
#>  8      8   6.02    17 <named list [3]>
#>  9      9   3.51    19 <named list [3]>
#> 10     10   1       21 <named list [3]>