Neighbor's masher

If you use Rstudio, the masher and spicer functions can help remind you which parameters go along with which ipa_brew flavor. The basic idea is to write spice(brew, with = spicer_<flavor>()) and mash(brew, with = masher_<flavor>()). Hitting the tab key with your curser inside the parentheses of masher_flavor()will create a drop-down menu that shows a list of the arguments that go along with your brew's flavor.

If you have no trouble remembering the parameters that go along with your brew's flavor, or if you just want your code to be more concise, you don't have to use the with argument. Instead, you can just specify parameter values directly using the ... argument in the mash and spice functions. In the examples below, both approaches are shown.

masher_nbrs(
  epsilon = 1e-08,
  nthread = NULL,
  fun_aggr_ctns = mean,
  fun_aggr_intg = medn_est,
  fun_aggr_catg = mode_est
)

Arguments

epsilon
Computed numbers (variable ranges) smaller than eps are treated as zero

nthread
Number of threads to use for parallelization. By default, for a dual-core machine, 2 threads are used. For any other machine n-1 cores are used so your machine doesn't freeze during a big computation. The maximum nr of threads are determined using omp_get_max_threads at C level.

fun_aggr_ctns
a function used to aggregate neighbors for continuous variables. If unspecified, the mean() function is used.

fun_aggr_intg
a function used to aggregate neighbors for integer values variables. If unspecified, the medn_est() function is used. This function returns the median of neighbor values, rounded to the nearest integer. medn_est_conserve() goes one step further and identifies which neighbor value is closest to the median, and returns that value. Both of these options can be helpful for integer valued columns if you want to make sure the imputed values do not contain impossible quantities, e.g. no. of children = 3/4.

fun_aggr_catg
a function used to aggregate neighbors for categorical variables. If unspecified, the mode_est() function is used.

Value

a list with input values that can be passed directly into mash, e.g mash(brew, with = masher_nbrs()) for a neighbors brew or mash(brew, with = masher_soft()) for a soft brew.

Examples

x1 = rnorm(100)
x2 = rnorm(100) + x1
x3 = rnorm(100) + x1 + x2

outcome = 0.5 * (x1 - x2 + x3)

data <- data.frame(x1=x1, x2=x2, x3=x3, outcome=outcome)

n_miss = 10

data[1:n_miss,'x1'] = NA
sft_brew <- brew_soft(data, outcome=outcome, bind_miss = FALSE)

# these two calls are equivalent
mash(sft_brew, with = masher_soft(bs = FALSE))

#> A brew to handle missing data using soft imputation. 
#> Data used for imputation (outcomes are wittheld): 
#>              x1         x2        x3
#>           <num>      <num>     <num>
#>   1:         NA  1.6155571  3.579510
#>   2:         NA  3.1508703  5.762276
#>   3:         NA -1.7713596 -2.425330
#>   4:         NA -0.8826988 -2.854109
#>   5:         NA -1.4809443 -2.610147
#>  ---                                
#>  96: -0.9222974 -1.5030021 -1.626649
#>  97: -1.0601146 -2.6531995 -1.962740
#>  98: -1.7871996 -2.2762422 -3.798112
#>  99: -1.0665518 -1.8192519 -2.568649
#> 100: -1.1298292 -0.4644504 -1.743016

mash(sft_brew, bs = FALSE)

#> A brew to handle missing data using soft imputation. 
#> Data used for imputation (outcomes are wittheld): 
#>              x1         x2        x3
#>           <num>      <num>     <num>
#>   1:         NA  1.6155571  3.579510
#>   2:         NA  3.1508703  5.762276
#>   3:         NA -1.7713596 -2.425330
#>   4:         NA -0.8826988 -2.854109
#>   5:         NA -1.4809443 -2.610147
#>  ---                                
#>  96: -0.9222974 -1.5030021 -1.626649
#>  97: -1.0601146 -2.6531995 -1.962740
#>  98: -1.7871996 -2.2762422 -3.798112
#>  99: -1.0665518 -1.8192519 -2.568649
#> 100: -1.1298292 -0.4644504 -1.743016

knn_brew <- brew_nbrs(data, outcome=outcome, bind_miss = TRUE) %>%

# these two calls are equivalent
mash(knn_brew, with = masher_nbrs(fun_aggr_ctns = median))
mash(knn_brew, fun_aggr_ctns = median)

#> A brew to handle missing data using k-nearest-neighbors. 
#> Data used for imputation (outcomes are wittheld): 
#>              x1         x2        x3 x1_missing
#>           <num>      <num>     <num>      <int>
#>   1:         NA  1.6155571  3.579510          1
#>   2:         NA  3.1508703  5.762276          1
#>   3:         NA -1.7713596 -2.425330          1
#>   4:         NA -0.8826988 -2.854109          1
#>   5:         NA -1.4809443 -2.610147          1
#>  ---                                           
#>  96: -0.9222974 -1.5030021 -1.626649          0
#>  97: -1.0601146 -2.6531995 -1.962740          0
#>  98: -1.7871996 -2.2762422 -3.798112          0
#>  99: -1.0665518 -1.8192519 -2.568649          0
#> 100: -1.1298292 -0.4644504 -1.743016          0