Soft masher

If you use Rstudio, the masher and spicer functions can help remind you which parameters go along with which ipa_brew flavor. The basic idea is to write spice(brew, with = spicer_<flavor>()) and mash(brew, with = masher_<flavor>()). Hitting the tab key with your curser inside the parentheses of masher_flavor()will create a drop-down menu that shows a list of the arguments that go along with your brew's flavor.

If you have no trouble remembering the parameters that go along with your brew's flavor, or if you just want your code to be more concise, you don't have to use the with argument. Instead, you can just specify parameter values directly using the ... argument in the mash and spice functions. In the examples below, both approaches are shown.

masher_soft(
  bs = TRUE,
  bs_maxit = 20,
  bs_thresh = 1e-09,
  bs_row.center = FALSE,
  bs_col.center = TRUE,
  bs_row.scale = FALSE,
  bs_col.scale = TRUE,
  si_type = "als",
  si_thresh = 1e-05,
  si_maxit = 100,
  si_final.svd = TRUE
)

Arguments

bs
a logical value. If TRUE, softImpute::biScale() is applied to data_ref or rbind(data_ref, data_new) prior to fitting softImpute models.

bs_maxit
an integer indicating the maximum number of iterations for the biScale algorithm.

bs_thresh
convergence threshold for the biScale algorithm.

bs_row.center
a logical value. If TRUE, row centering will be performed. If FALSE (default), then nothing is done.

bs_col.center
a logical value. If TRUE (default), column centering will be performed. If FALSE, then nothing is done.

bs_row.scale
a logical value. If TRUE, row scaling will be performed. If FALSE (default), then nothing is done.

bs_col.scale
a logical value. If TRUE (default), column scaling will be performed. If FALSE, then nothing is done.

si_type
two algorithms are implemented, type="svd" or the default type="als". The "svd" algorithm repeatedly computes the svd of the completed matrix, and soft thresholds its singular values. Each new soft-thresholded svd is used to re-impute the missing entries. For large matrices of class "Incomplete", the svd is achieved by an efficient form of alternating orthogonal ridge regression. The "als" algorithm uses this same alternating ridge regression, but updates the imputation at each step, leading to quite substantial speedups in some cases. The "als" approach does not currently have the same theoretical convergence guarantees as the "svd" approach.

si_thresh
convergence threshold for the softImpute algorithm, measured as the relative change in the Frobenius norm between two successive estimates.

si_maxit
maximum number of iterations for the softImpute algorithm.

si_final.svd
only applicable to si_type = "als". The alternating ridge-regressions do not lead to exact zeros. With the default final.svd = TRUE, at the final iteration, a one step unregularized iteration is performed, followed by soft-thresholding of the singular values, leading to hard zeros.

Value

a list with input values that can be passed directly into mash, e.g mash(brew, with = masher_nbrs()) for a neighbors brew or mash(brew, with = masher_soft()) for a soft brew.

Examples

x1 = rnorm(100)
x2 = rnorm(100) + x1
x3 = rnorm(100) + x1 + x2

outcome = 0.5 * (x1 - x2 + x3)

data <- data.frame(x1=x1, x2=x2, x3=x3, outcome=outcome)

n_miss = 10

data[1:n_miss,'x1'] = NA
sft_brew <- brew_soft(data, outcome=outcome, bind_miss = FALSE)

# these two calls are equivalent
mash(sft_brew, with = masher_soft(bs = FALSE))

#> A brew to handle missing data using soft imputation. 
#> Data used for imputation (outcomes are wittheld): 
#>               x1          x2         x3
#>            <num>       <num>      <num>
#>   1:          NA -1.51265576 -1.3208307
#>   2:          NA  1.08698578  1.7135815
#>   3:          NA -0.40352826 -0.7160933
#>   4:          NA -0.03375898  0.2829920
#>   5:          NA -1.74378020 -3.4624022
#>  ---                                   
#>  96:  0.19962995 -1.09237095 -1.8065284
#>  97:  1.71290679 -0.01762329  1.4226881
#>  98:  0.55273636  1.68516114  2.9889841
#>  99: -0.86873581  0.08666271 -0.8589293
#> 100:  0.09863716 -0.61537112 -0.3696324

mash(sft_brew, bs = FALSE)

#> A brew to handle missing data using soft imputation. 
#> Data used for imputation (outcomes are wittheld): 
#>               x1          x2         x3
#>            <num>       <num>      <num>
#>   1:          NA -1.51265576 -1.3208307
#>   2:          NA  1.08698578  1.7135815
#>   3:          NA -0.40352826 -0.7160933
#>   4:          NA -0.03375898  0.2829920
#>   5:          NA -1.74378020 -3.4624022
#>  ---                                   
#>  96:  0.19962995 -1.09237095 -1.8065284
#>  97:  1.71290679 -0.01762329  1.4226881
#>  98:  0.55273636  1.68516114  2.9889841
#>  99: -0.86873581  0.08666271 -0.8589293
#> 100:  0.09863716 -0.61537112 -0.3696324

knn_brew <- brew_nbrs(data, outcome=outcome, bind_miss = TRUE) %>%

# these two calls are equivalent
mash(knn_brew, with = masher_nbrs(fun_aggr_ctns = median))
mash(knn_brew, fun_aggr_ctns = median)

#> A brew to handle missing data using k-nearest-neighbors. 
#> Data used for imputation (outcomes are wittheld): 
#>               x1          x2         x3 x1_missing
#>            <num>       <num>      <num>      <int>
#>   1:          NA -1.51265576 -1.3208307          1
#>   2:          NA  1.08698578  1.7135815          1
#>   3:          NA -0.40352826 -0.7160933          1
#>   4:          NA -0.03375898  0.2829920          1
#>   5:          NA -1.74378020 -3.4624022          1
#>  ---                                              
#>  96:  0.19962995 -1.09237095 -1.8065284          0
#>  97:  1.71290679 -0.01762329  1.4226881          0
#>  98:  0.55273636  1.68516114  2.9889841          0
#>  99: -0.86873581  0.08666271 -0.8589293          0
#> 100:  0.09863716 -0.61537112 -0.3696324          0