Sip a brew

Sip a brew

sip(
  brew,
  from = "training",
  data_complete,
  fun_ctns_error = yardstick::rsq_vec,
  fun_intg_error = yardstick::rsq_vec,
  fun_bnry_error = yardstick::kap_vec,
  fun_catg_error = yardstick::kap_vec
)

Arguments

brew
an ipa_brew object.

from
column name for sipping. Valid options are training and testing. Inputs can be quoted (e.g. 'training') or unquoted (e.g. training).

data_complete
a data frame containing the 'true' values that were 'missing'.

fun_ctns_error
a function that will evaluate errors for continuous variables. Continuous variables have type double. Default is to use R-squared (see yardstick::rsq()).

fun_intg_error
a function that will evaluate errors for integer valued variables. Default is to use R-squared (see yardstick::rsq()).

fun_bnry_error
a function that will evaluate errors for binary variables (i.e., factors with 2 levels). Default is to use kappa agreement (see yardstick::kap()).

fun_catg_error
a function that will evaluate errors for categorical variables (i.e., factors with >2 levels). Default is to use kappa agreement (see yardstick::kap()).

Value

an ipa_brew object with a new column added to the wort. The new column contains a list of tibble::tibble()s with columns variable, type, and score. The score column comprises output from the error functions.

Examples

data("diabetes")

df_miss <- diabetes$missing
df_cplt <- diabetes$complete

sft_brew <- brew_soft(df_miss, outcome = diabetes) %>%
  mash() %>%
  stir() %>%
  ferment() %>%
  bottle(type = 'tibble')

sip(sft_brew, from = training, data_complete = df_cplt)

#> A brew to handle missing data using soft imputation. 
#>     impute   pars training training_score
#>      <int> <list>   <list>         <list>
#>  1:      1 <list> <tbl_df>       <tbl_df>
#>  2:      2 <list> <tbl_df>       <tbl_df>
#>  3:      3 <list> <tbl_df>       <tbl_df>
#>  4:      4 <list> <tbl_df>       <tbl_df>
#>  5:      5 <list> <tbl_df>       <tbl_df>
#>  6:      6 <list> <tbl_df>       <tbl_df>
#>  7:      7 <list> <tbl_df>       <tbl_df>
#>  8:      8 <list> <tbl_df>       <tbl_df>
#>  9:      9 <list> <tbl_df>       <tbl_df>
#> 10:     10 <list> <tbl_df>       <tbl_df>