Stir your brew

After setting your primary and secondary imputation parameters (i.e., running spice and mash), you're ready to fit the imputation model(s) that will impute missing values in your training data. This is what the stir function does.

stir(brew, timer = FALSE)

Arguments

brew
an ipa_brew object.

timer
a logical value. If TRUE, then the amount of time it takes to fit the imputation models will be tracked and saved as an attribute of the resulting ipa_brew object.

Value

an ipa_brew object with imputed values added to the wort

Examples

x1 = rnorm(100)
x2 = rnorm(100) + x1
x3 = rnorm(100) + x1 + x2

outcome = 0.5 * (x1 - x2 + x3)

n_miss = 10
x1[1:n_miss] <- NA

data <- data.frame(x1=x1, x2=x2, x3=x3, outcome=outcome)

sft_brew <- brew_soft(data, outcome=outcome, bind_miss = FALSE)
sft_brew <- mash(sft_brew, with = masher_soft(bs = TRUE))
sft_brew <- stir(sft_brew, timer = TRUE)
sft_brew

#> A brew to handle missing data using soft imputation. 
#>     impute   lambda rank_max rank_fit    fit iv_training
#>      <int>    <num>    <int>    <int> <list>      <list>
#>  1:      1 1.200000        2        2 <list>      <list>
#>  2:      2 1.177778        2        2 <list>      <list>
#>  3:      3 1.155556        2        2 <list>      <list>
#>  4:      4 1.133333        2        2 <list>      <list>
#>  5:      5 1.111111        2        2 <list>      <list>
#>  6:      6 1.088889        2        2 <list>      <list>
#>  7:      7 1.066667        2        2 <list>      <list>
#>  8:      8 1.044444        2        2 <list>      <list>
#>  9:      9 1.022222        2        2 <list>      <list>
#> 10:     10 1.000000        2        2 <list>      <list>